+++ Koronavilkku/Koronavilkku/UI/Styles/Fonts.swift
@@ -8,9 +8,8 @@ extension UIFont {
     static var heading4 = useFont(name: "SourceSansPro-SemiBold", ofSize: 18, withStyle: .headline)
     static var heading5 = useFont(name: "SourceSansPro-SemiBold", ofSize: 13, withStyle: .subheadline)
     static var labelPrimary = useFont(name: "SourceSansPro-Bold", ofSize: 17, withStyle: .headline)
-    static var labelSecondary = useFont(name: "SourceSansPro-Bold", ofSize: 13, withStyle: .caption1)
-    static var labelTertiary = useFont(name: "SourceSansPro-Regular", ofSize: 13, withStyle: .caption1)
-    static var labelFooter = useFont(name: "SourceSansPro-Regular", ofSize: 14, withStyle: .caption1)
+    static var labelSecondary = useFont(name: "SourceSansPro-Bold", ofSize: 14, withStyle: .caption1)
+    static var labelTertiary = useFont(name: "SourceSansPro-Regular", ofSize: 14, withStyle: .caption1)
     static var coronaCode = useFont(name: "SourceSansPro-Regular", ofSize: 24, withStyle: .subheadline)
     static var bodyLarge = useFont(name: "SourceSansPro-Regular", ofSize: 18, withStyle: .body)
     static var bodySmall = useFont(name: "SourceSansPro-Regular", ofSize: 15, withStyle: .body)
